requirements.txt
PyYAML
google-cloud-bigquery
google-cloud-aiplatform
feedparser
tweepy
google-generativeai
requests
pydantic
Pillow
